ArticleController Crud functions added

+++ src/main/java/de/supercode/tastaturen_shop/services/ArticleService.java
@@ -2,7 +2,11 @@ package de.supercode.tastaturen_shop.services;
 
 import de.supercode.tastaturen_shop.entities.article.Article;
 import de.supercode.tastaturen_shop.repositories.ArticleRepository;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ArticleService {
@@ -11,11 +15,14 @@ public class ArticleService {
     public ArticleService(ArticleRepository articleRepository) {
         this.articleRepository = articleRepository;
     }
-    public void createArticle(Article article){
-        this.articleRepository.save(article);
+    public Article createArticle(Article article){
+        return this.articleRepository.save(article);
     }
     public Article getArticle(long id){
         return articleRepository.findById(id).orElse(null);
+    }
+    public List<Article> getAllArticles(){
+        return articleRepository.findAll();
     }
     public Article updateArticle(long id, Article article) {
         if (article == null) return null;
@@ -28,7 +35,12 @@ public class ArticleService {
         toUpdateArticle.setStatus(article.isStatus());
         return articleRepository.save(toUpdateArticle);
     }
-    public void deleteArticle(long id){
-       articleRepository.deleteById(id);
+    public Boolean deleteArticle(long id){
+        Optional<Article> toDeleteArticle = articleRepository.findById(id);
+        if (toDeleteArticle.isEmpty()) return false;
+        else {
+            articleRepository.deleteById(id);
+            return true;
+        }
     }
 }
